refactor: DistributedLock Repository -> Manager 폴더명 변경 (#320)
