+++ ttoklip-infrastructure/src/main/java/com/infrastructure/distribution/manager/DistributedLockManager.java
@@ -1,4 +1,4 @@
-package com.infrastructure.distribution.repository;
+package com.infrastructure.distribution.manager;
 
 public interface DistributedLockManager {
     boolean tryLock(String key, long waitTime, long leaseTime);
