+++ ttoklip-infrastructure/src/main/java/com/infrastructure/distribution/manager/LocalLockManager.java
@@ -1,4 +1,4 @@
-package com.infrastructure.distribution.repository;
+package com.infrastructure.distribution.manager;
 
 import java.util.Map;
 import java.util.Set;
