+++ ttoklip-infrastructure/src/main/java/com/infrastructure/distribution/service/LockSynchronizationService.java
@@ -1,7 +1,7 @@
 package com.infrastructure.distribution.service;
 
-import com.infrastructure.distribution.repository.LocalLockManager;
-import com.infrastructure.distribution.repository.RedissonLockManager;
+import com.infrastructure.distribution.manager.LocalLockManager;
+import com.infrastructure.distribution.manager.RedissonLockManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
